tests: Remove May15 test
This test is no longer relevant.

It was introduced in 681535dc87676747c6d354deee370bf9e18fe235 to check
the switch to 1MB blocks after the BDB too-many-locks issue back in
2013. The switching code has been long since removed.

It also needs a specific data file that is hard to find. I've verified
in #6320 that it still passes, however I think there is zero reason to
keep it.

Closes #6320.