Bugfix: the end of a reorged chain is invalid when connect fails
Introduced in 1ca2f802aed265feb8c3fef3c6b7e8bef3c1c68a

When an invalid block is found during a reorg, we know the last of the blocks in
the was-to-be-connected chain is invalid, but not necessarily the first. As
vpIndexToConnect is ordered in decreasing height, the end of the reorg is the
front of the vector, and not the back.

This only affected the warning system.