[addrman] Fix new table bucketing during unserialization
An addrman entry can appear in up to 8 new table buckets. We store this
entry->bucket indexing during shutdown so that on restart we can restore
the entries to their correct buckets.

Commit 45a0d2d16e50c0c18d1453d282c106699098979e broke the
deserialization code so that each entry could only be put in up to one
new bucket. Fix that.